Feat/#34: 패키지 구조 변경 및 리팩토링
