+++ src/main/java/com/example/cluvrbatch/CluvrBatchApplication.java
@@ -1,10 +1,12 @@
 package com.example.cluvrbatch;
 
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableBatchProcessing
 @EnableScheduling
 public class CluvrBatchApplication {
