+++ src/main/java/com/example/cluvrbatch/job/cloverlog/steps/redisToDbStep/CloverLogItemWriter.java
@@ -1,4 +1,4 @@
-package com.example.cluvrbatch.job.cloverlog;
+package com.example.cluvrbatch.job.cloverlog.steps.redisToDbStep;
 
 import lombok.RequiredArgsConstructor;
