+++ src/main/java/com/example/cluvrbatch/job/gemlog/steps/redisToDbStep/GemLogItemReader.java
@@ -1,4 +1,4 @@
-package com.example.cluvrbatch.job.cloverlog;
+package com.example.cluvrbatch.job.gemlog.steps.redisToDbStep;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -11,28 +11,28 @@ import org.springframework.batch.item.ItemReader;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
-import com.example.cluvrbatch.job.cloverlog.dto.CloverEventResponseDto;
 import com.example.cluvrbatch.job.enums.RedisKey;
+import com.example.cluvrbatch.job.gemlog.dto.GemEventResponseDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Component
 @RequiredArgsConstructor
-public class CloverLogItemReader implements ItemReader<CloverEventResponseDto> {
+public class GemLogItemReader implements ItemReader<GemEventResponseDto> {
 
 	private final RedisTemplate<String, String> redisTemplate;
 	private final ObjectMapper objectMapper;
 
-	private Queue<CloverEventResponseDto> cached = new LinkedList<>();
+	private Queue<GemEventResponseDto> cached = new LinkedList<>();
 
 	@Override
-	public CloverEventResponseDto read() throws Exception {
+	public GemEventResponseDto read() throws Exception {
 		if (cached.isEmpty()) {
-			Set<String> keys = redisTemplate.keys(RedisKey.CLOVER_LOG.getKey() + "*");
+			Set<String> keys = redisTemplate.keys(RedisKey.GEM_LOG.getKey() + "*");
 
 			for (String key : keys) {
 				List<String> logs = redisTemplate.opsForList().range(key, 0, -1);
 				for (String json : logs) {
-					cached.add(objectMapper.readValue(json, CloverEventResponseDto.class));
+					cached.add(objectMapper.readValue(json, GemEventResponseDto.class));
 				}
 				redisTemplate.delete(key); // 삭제
 			}
