--- src/main/java/com/example/cluvrbatch/job/useractivity/BoardStatJobService.java
@@ -1,24 +1,0 @@
-package com.example.cluvrbatch.job.useractivity;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.JobParametersBuilder;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.stereotype.Component;
-
-@RequiredArgsConstructor
-@Component
-public class BoardStatJobService {
-	private final JobLauncher jobLauncher;
-	private final Job boardLogJob;
-
-	public void runJob() throws Exception {
-		JobParameters jobParameters = new JobParametersBuilder()
-			.addLong("time", System.currentTimeMillis())
-			.toJobParameters();
-
-		jobLauncher.run(boardLogJob, jobParameters);
-	}
-}
