+++ src/main/java/com/example/cluvrbatch/job/useractivity/steps/redisToDbStep/BoardStatItemWriter.java
@@ -1,4 +1,4 @@
-package com.example.cluvrbatch.job.useractivity;
+package com.example.cluvrbatch.job.useractivity.steps.redisToDbStep;
 
 import lombok.RequiredArgsConstructor;
